explicit conversions of data types added
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/filevector@242 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ fvlib/filevector.cpp
@@ -121,13 +121,17 @@ void filevector::set_cachesizeMb( unsigned long int cachesizeMb )
 	if (cache_size_nvars<1) {
 		message("attempting to set cache size to 1 var (%f Mb)\n",
 		         (float) data_type.nobservations*data_type.bytes_per_record/(1024.*1024.));
-		cache_size_Mb = ceil((float) data_type.nobservations*data_type.bytes_per_record/(1024.*1024.));
+		cache_size_Mb = (long unsigned int) ceil(
+				(float) data_type.nobservations*data_type.bytes_per_record/(1024.*1024.)
+				);
 		cache_size_nvars = 1;
 	} else if (cache_size_nvars>data_type.nvariables) {
 		message("attempting to cache all the data (%u variables, %f Mb)\n",
 			 data_type.nvariables,
 		         (float) data_type.nvariables*data_type.nobservations*data_type.bytes_per_record/(1024.*1024.));
-		cache_size_Mb = ceil((float) data_type.nvariables*data_type.nobservations*data_type.bytes_per_record/(1024.*1024.));
+		cache_size_Mb = (long unsigned int) ceil(
+				(float) data_type.nvariables*data_type.nobservations*data_type.bytes_per_record/(1024.*1024.)
+				);
 		cache_size_nvars = data_type.nvariables;
 	} else {
 		message("attempting to cache specified amount of data (%u variables, %f Mb)\n",
